Refactor into API.ts
